pyproject.toml
# PyProject.toml to be used as the base for the generated SDK

[project]
name = "python-zendesk"
version = "0.1.15"
description = "Python Client SDK for Zendesk Support API."
authors = [{ name = "Ben Meyer-Meisel" }, { name = "botBrains"}, ]
readme = "README.md"
requires-python = ">=3.9.2"
dependencies = [
    "httpcore >=1.0.9",
    "httpx >=0.28.1",
    "jsonpath-python >=1.0.6",
    "pydantic >=2.11.2",
    "pyyaml (>=6.0.2,<7.0.0)",
]
project_urls = { Homepage = "https://github.com/botbrains-io/python-zendesk", Repository = "https://github.com/botbrains-io/python-zendesk.git", Documentation = "https://github.com/botbrains-io/python-zendesk#readme" }

[tool.poetry]
name = "python-zendesk"
version = "0.1.15"
description = "Python Client SDK for Zendesk Support API."
authors = ["Ben Meyer-Meisel"]
readme = "README.md"
packages = [{include = "zendesk", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9.2"
httpcore = ">=1.0.9"
httpx = ">=0.28.1"
jsonpath-python = ">=1.0.6"
pydantic = ">=2.11.2"

[tool.poetry.group.dev.dependencies]
mypy = "==1.15.0"
pylint = "==3.2.3"
pyyaml = ">=6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]

[tool.mypy]
disable_error_code = "misc"

[[tool.mypy.overrides]]
module = "typing_inspect"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jsonpath"
ignore_missing_imports = true

[tool.pyright]
venvPath = "."
venv = ".venv"

